requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-multiselectfield==0.1.12
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.5
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pipreqs==0.4.11
protobuf==4.22.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
yarg==0.1.9
